Add scaler class
